Restrict GITHUB_TOKEN permissions
This commit restricts what GITHUB_TOKEN can do.

As per [this blog post](https://github.blog/changelog/2021-04-20-github-actions-control-permissions-for-github_token/), all other permissions are set to `none` by default if not specified under the jobs permissions.

`contents` is used to access the repository tree, while `issues` is used to (of course) access issues.

If either of these permissions are *not* `write`, the workflow fails. If anything requires neither of these permissions, it will be unable to access them.